Cleanups and beginning to reformat (#1)
* Ugly mechanical whitespace to space

* ASCIIfy this mess! (Intend to revert)

I started working to reformat some paragraphs to a textwidth of 78
(which is pretty common for markdown), but the random UTF-8 where it
wasn't needed was kind of in the way. I ruthlessly iconv'd this thing to
ASCII using some replaceable symbols like [-1/2] for the ½ fraction and
[x] for the ×.

These are things I can pretty trivially type (which means I can search
and replace to revert them later) but … I think I want them gone while I
am working with reformatting paragraphs and the like in a very rough
form. Putting them back once the document begins to take shape is
literally just a search and replace.

What won't be so trivial to undo is replacing all the various lengths of
-, –, —, etc with a basic ASCII -. I get that the big long lines look
better in tables and whatnot and maybe when the time comes they should
be put back … but the tables are basically broken AF anyway and while I
probably brok them more, I'll burn that bridge when I come to it. 😅

* A little more reflowing