Big improvement to license.md
Put the actual text from the first page of the SRD into the license file
so that the attribution clause required is front and center.

Added a disclaimer that we are adding anything new here that would
someohow cause this to have some new Copyright (not sure everyone here
wants a real name attached to the project anyway…)

Reformatted the legalcode of the license. Could've dropped actual HTML
in here, but initially thought why not Markdown it? Because standard
Markdown doesn't have the ability to do more complex ordered lists.

Fine, grabbed a formatted version from a Debian copyright file and
verified the version was the same.