Format ability blocks in ch 12
Had to convert these to kramdown/markuru format because *sigh* pandoc
doesn't support attribute blocks on tables yet! That's really quite
annoying because pandoc has more features. We can still use it to
generate ePub if we want to, but … not from Markdown. Asciidoc maybe?

If a later version of pandoc has this, let me know—2.17 doesn't, at
least for the Markdown reader.