Oneline the ability block numbers for creatures
Okay, there's a process here:

1. Get the ability blocks into the format:
   ## (±##) ## (±##) ## (±##) ## (±##) ## (±##) ## (±##)
2. Apply a couple seds to put it into correct table format
3. Remove any of these stray fenced code blocks that are everywhere

This is the creatures version of #1 above, done as a proof of concept. I
have already designed the s/// commands to do #2 but … let's get this
far with the monsters first. Then I'll start writing the macros to fix
the fenced code blocks. A few more commits to go and I'll push these out
together.

I'm debating whether or not to squash them before I do. My sense is that
I won't, and I'll explain why for anyone who stumbles across this commit
log: This is a considerable amount of work! When some loser suit at
WotC/Hasbro starts talking about how this game belongs to them, as if it
was something they worked hard for and us "freeloaders" are just using
for free like it was nothing … it isn't.

It would probably actually be easier to recreate the content than
reformat theirs—but people value the promise they made that this stuff
belonged to the world now and have depended on that to make the D&D
ecosystem larger. Easy, huh?

Yeah, easy. I'ma make me a bandit:

        Bandit

        medium humanoid (any race), any non-lawful alignment

        Armor Class: 12 (light armor)
        HP: 9 (2d8)
        Speed: 30ft

        ----------------------------
        STR  DEX  CON  INT  WIS  CHA
        +1   +1   +1   -1   +0   +0
        ----------------------------

        Senses: Passive perception 10

        Languages: Native for background, at least some common typical

        Challenge: 1/4 (50 xp)

        Actions

        Sword. +3 to hit, 5ft reach, hit: 6 (d6 + 3) slashing
        Crossbow. (Ranged) +3 to hit, 60-300ft range,
                hit: 7 (d8 +3) piercing
        Knife. 3ft reach, hit 3 (d4 + 1) slashing/piercing

        Bandits operate in bands that will raid or lie in ambush for
        those with valuables to steal. They needn't be evil, since
        desperation leads to desperate actions. Regardless of their
        native language, they'll typically speak enough common to
        communicate their demands. Banditry is a nervous profession
        since any bandit taken into custody by local authorities faces
        execution within 48 hours.

Good as the 5E bandit? Probably. I just designed common humanoid aiming
for level zero and a half, fairly tough/fast, but not too bright, and
armed well enough to be a threat to other common folk. Especially if
faced with a band of bandits. And I came up with it in less time than it
took me to look up the 5E stat block for such a creature.

But my bandit is gonna be different than yours, etc. The whole idea was
that WotC put *their* bandit under a free license so we'd all agree to
use that one, or something close to it. And in exchange for that they
got to be the biggest game out there. If they want to give that up with
6E, it's their loss—but they can't take away 5E.

And that's why we're doing all this work.