Some class cleanup
I removed a bunch of the code blocks that got inserted into this, and
did the reformat on a couple classes. Many of them have a table that's
not going to be possible to represent as in the PDF with pure Markdown.
I finished turning Bard's table into the closest to whatever I can until
I decide how to handle that kind of table. Again this seems like a job
for Jekyll and data files. Jekyll can generate the requisite HTML from a
data source, and the same data source could be used to generate
something with LaTeX for print if desired. Wouldn't even be hard (if you
know LaTeX … which I don't really.)

Sleep now, figure out what to do with that later.